style: linter fix
