+++ src/pages/start-game/components/index.ts
@@ -1,0 +1,3 @@
+export { PaperGame } from './styled';
+export { Rules } from './rules';
+export { Progress } from './progress';
