+++ src/pages/start-game/components/rules.tsx
@@ -1,7 +1,7 @@
 import { Alert, Stack } from '@mui/material';
 
 export const Rules = () => (
-  <Stack sx={{ width: '100%' }} spacing={2} mt={4}>
+  <Stack spacing={2} mt={4} width="100%">
     <Alert severity="info" onClose={() => {}}>
       Тут будут правила игры
     </Alert>
