+++ src/pages/start-game/start-game.tsx
@@ -1,10 +1,8 @@
 import { useState, useEffect } from 'react';
-import type { FC } from 'react';
-import { Typography, Box, Skeleton, Paper, Button } from '@mui/material';
-import { Progress } from 'pages/start-game/progress';
-import { Rules } from 'pages/start-game/rules';
+import { Typography, Box, Skeleton, Button } from '@mui/material';
+import { Progress, Rules, PaperGame } from './components';
 
-export const StartGamePage: FC = () => {
+export const StartGamePage = () => {
   const [progress, setProgress] = useState(10);
 
   useEffect(() => {
@@ -29,22 +27,12 @@ export const StartGamePage: FC = () => {
         </Box>
       ) : (
         <Box>
-          <Paper
-            sx={{
-              width: '100%',
-              height: '400px',
-              backgroundColor: '#e0e0e0',
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-              flexDirection: 'column',
-            }}
-          >
+          <PaperGame>
             <Typography variant="h2">Все готово!</Typography>
             <Button size="large" variant="contained">
               Играть
             </Button>
-          </Paper>
+          </PaperGame>
         </Box>
       )}
       <Rules />
